Algod shared publishers

+++ src/main/kotlin/se/newton/stockpriceclient/controller/rest/models/ShortBlockSummary.kt
@@ -7,7 +7,7 @@ package se.newton.stockpriceclient.controller.rest.models
  */
 data class ShortBlockSummary(
 	val net: String,
-	val round: Long,
+	val round: Int,
 	val transactions: Int,
 ) {
 }
